refactor: download exported admin json
