+++ libs/frontend/application/admin/src/use-cases/export-data/ExportAdminData.action.tsx
@@ -5,6 +5,7 @@ import type { IExportDto } from '@codelab/shared/abstract/core'
 
 export const exportAdminDataAction = async (data: IExportDto) => {
   const apiClient = await getAuthenticatedApiClient()
+  const response = await apiClient.post('admin/export', data)
 
-  await apiClient.post('admin/export', data)
+  return response.data
 }
