+++ libs/frontend/application/admin/src/use-cases/export-data/ExportAdminDataModal.tsx
@@ -1,20 +1,31 @@
 'use client'
 
 import { MODEL_ACTION } from '@codelab/frontend/abstract/types'
-import { createFormErrorNotificationHandler } from '@codelab/frontend/shared/utils'
+import {
+  createFormErrorNotificationHandler,
+  downloadJsonAsFile,
+} from '@codelab/frontend/shared/utils'
 import { ModalForm } from '@codelab/frontend-presentation-components-form'
 import {
   exportDtoDefault,
   ExportDtoSchema,
   type IExportDto,
 } from '@codelab/shared/abstract/core'
-import React from 'react'
+import React, { useCallback } from 'react'
 import { AutoFields } from 'uniforms-antd'
 import { exportAdminDataAction } from './ExportAdminData.action'
 import { useExportAdminDataModal } from './ExportAdminDataModal.state'
 
 export const ExportAdminDataModal = () => {
   const exportDataModal = useExportAdminDataModal()
+
+  const onSubmitHandler = useCallback(async (data: IExportDto) => {
+    const exportedData = await exportAdminDataAction(data)
+
+    if (exportedData) {
+      downloadJsonAsFile('export.json', exportedData)
+    }
+  }, [])
 
   return (
     <ModalForm.Modal
@@ -24,7 +35,7 @@ export const ExportAdminDataModal = () => {
     >
       <ModalForm.Form<IExportDto>
         model={exportDtoDefault}
-        onSubmit={exportAdminDataAction}
+        onSubmit={onSubmitHandler}
         onSubmitError={createFormErrorNotificationHandler({
           title: 'Error while exporting data',
         })}
